# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Associati digilab",
        "description": "Api per la gestione degli associati del digilab",
        "version": "0.1"
    },
    "paths": {
...
        "/api/lista-corsi": {
            "get": {
                "tags": [
                    "gestione corsi"
                ],
                "summary": "Recupera la lista dei corsi",
                "operationId": "4a9bbe2eafbb67010473c5fb37e9ed07",
                "responses": {
                    "200": {
                        "description": "lista corsi"
                    },
                    "500": {
                        "description": "Errore nel recupero della lista degli associati"
                    }
                }
            }
        },
        "/api/nuovo-corso": {
            "post": {
                "tags": [
                    "gestione corsi"
                ],
                "summary": "Aggiunge un nuovo corso",
                "description": "inserisce i dati di un nuovo corso nel database",
                "operationId": "e7e426da3e2ba69403fd902717135362",
                "requestBody": {
                    "description": "dati corso",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titolo",
                                    "descrizione",
                                    "data_inizio",
                                    "data_fine",
                                    "costo"
                                ],
                                "properties": {
                                    "titolo": {
                                        "type": "string",
                                        "format": "text"
                                    },
                                    "descrizione": {
                                        "type": "string",
                                        "format": "text"
                                    },
                                    "data_inizio": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "data_fine": {
                                        "type": "string",
                                        "format": "text"
                                    },
                                    "costo": {
                                        "type": "string",
                                        "format": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "corso aggiunto"
                    },
                    "204": {
                        "description": "dati mancanti o non validi"
                    },
                    "500": {
                        "description": "Errore nell'inserimento del nuovo corso"
                    }
                }
            }
        },
        "/api/elimina-corso/{id}": {
            "delete": {
                "tags": [
                    "gestione corsi"
                ],
                "summary": "Elimina un corso tramite il suo id",
                "operationId": "694e59018908326a1feb7905062876d8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id del corso da eliminare",
                        "required": true,
                        "schema": {
                            "description": "id corso",
                            "type": "string",
                            "minimum": 16
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "corso eliminato"
                    },
                    "404": {
                        "description": "corso non trovato"
                    }
                }
            }
        },
        "/api/corso/{id}": {
            "get": {
                "tags": [
                    "gestione corsi"
                ],
                "summary": "Recupera un corso tramite il suo id",
                "operationId": "14359d0498a28e47a4542aeec9e32a7e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id del corso da recuperare",
                        "required": true,
                        "schema": {
                            "description": "id corso",
                            "type": "string",
                            "minimum": 16
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "dati del corso"
                    },
                    "404": {
                        "description": "corso non trovato"
                    },
                    "500": {
                        "description": "errore nella richiesta"
                    }
                }
            }
        },
        "/api/modifica-corso/{id}": {
            "put": {
                "tags": [
                    "gestione corsi"
                ],
                "summary": "Modifica corso",
                "description": "modifica i dati del corso nel database",
                "operationId": "2837451d1681fba210db0f21ec708179",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id del corso da recuperare",
                        "required": true,
                        "schema": {
                            "description": "id corso",
                            "type": "string",
                            "minimum": 16
                        }
                    }
                ],
                "requestBody": {
                    "description": "dati corso",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "titolo",
                                    "descrizione",
                                    "data_inizio",
                                    "data_fine",
                                    "costo"
                                ],
                                "properties": {
                                    "titolo": {
                                        "type": "string",
                                        "format": "text"
                                    },
                                    "descrizione": {
                                        "type": "string",
                                        "format": "text"
                                    },
                                    "data_inizio": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "data_fine": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "costo": {
                                        "type": "string",
                                        "format": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "corso aggiunto"
                    },
                    "204": {
                        "description": "dati corso mancanti"
                    },
                    "500": {
                        "description": "Errore nell'aggiornamento del nuovo corso"
                    }
                }
            }
        },
...
    },
...
}